remove more apis. rename command_queue_commands to host_runtime_commands. move things around
