+++ tt_metal/impl/buffers/circular_buffer.cpp
@@ -12,13 +12,6 @@
 #include <device.hpp>
 #include <command_queue.hpp>
 
-namespace {
-
-inline void GetBufferAddress(const tt::tt_metal::Buffer* buffer, uint32_t* address_on_host) {
-    GetBufferAddr(address_on_host, buffer, false);
-}
-
-}  // namespace
 namespace tt {
 
 namespace tt_metal {
@@ -129,9 +122,7 @@ uint32_t CircularBuffer::address() const {
     return this->globally_allocated() ? globally_allocated_address_ : locally_allocated_address_.value();
 }
 
-void CircularBuffer::assign_global_address() {
-    GetBufferAddress(config_.shadow_global_buffer, &globally_allocated_address_);
-}
+void CircularBuffer::assign_global_address() { globally_allocated_address_ = config_.shadow_global_buffer->address(); }
 
 void CircularBuffer::set_global_circular_buffer(const v1::experimental::GlobalCircularBuffer& global_circular_buffer) {
     TT_FATAL(
