+++ tt_metal/tt_metal.cpp
@@ -241,6 +241,29 @@ inline void SetRuntimeArgsImpl(
     }
 }
 
+void SetRuntimeArgsImpl(
+    const std::shared_ptr<Kernel>& kernel,
+    const CoreCoord& core_coord,
+    std::shared_ptr<RuntimeArgs> runtime_args_ptr,
+    bool blocking) {
+    std::vector<uint32_t> resolved_runtime_args = {};
+    resolved_runtime_args.reserve(runtime_args_ptr->size());
+
+    for (const auto& arg : *(runtime_args_ptr)) {
+        std::visit(
+            [&resolved_runtime_args](auto&& a) {
+                using T = std::decay_t<decltype(a)>;
+                if constexpr (std::is_same_v<T, Buffer*>) {
+                    resolved_runtime_args.push_back(a->address());
+                } else {
+                    resolved_runtime_args.push_back(a);
+                }
+            },
+            arg);
+    }
+    kernel->set_runtime_args(core_coord, resolved_runtime_args);
+}
+
 inline void SetRuntimeArgsImpl(
     const std::shared_ptr<Kernel> kernel,
     const std::variant<CoreCoord, CoreRange, CoreRangeSet>& core_spec,
@@ -251,18 +274,18 @@ inline void SetRuntimeArgsImpl(
         [&](auto&& core_spec) {
             using T = std::decay_t<decltype(core_spec)>;
             if constexpr (std::is_same_v<T, CoreCoord>) {
-                SetRuntimeArgs(kernel, core_spec, runtime_args, blocking);
+                SetRuntimeArgsImpl(kernel, core_spec, runtime_args, blocking);
             } else if constexpr (std::is_same_v<T, CoreRange>) {
                 for (auto x = core_spec.start_coord.x; x <= core_spec.end_coord.x; x++) {
                     for (auto y = core_spec.start_coord.y; y <= core_spec.end_coord.y; y++) {
-                        SetRuntimeArgs(kernel, CoreCoord(x, y), runtime_args, blocking);
+                        SetRuntimeArgsImpl(kernel, CoreCoord(x, y), runtime_args, blocking);
                     }
                 }
             } else if constexpr (std::is_same_v<T, CoreRangeSet>) {
                 for (const auto& core_range : core_spec.ranges()) {
                     for (auto x = core_range.start_coord.x; x <= core_range.end_coord.x; x++) {
                         for (auto y = core_range.start_coord.y; y <= core_range.end_coord.y; y++) {
-                            SetRuntimeArgs(kernel, CoreCoord(x, y), runtime_args, blocking);
+                            SetRuntimeArgsImpl(kernel, CoreCoord(x, y), runtime_args, blocking);
                         }
                     }
                 }
@@ -278,7 +301,7 @@ inline void SetRuntimeArgsImpl(
     bool blocking) {
     // SetRuntimeArgs API for Async CQ Mode (support vector of runtime args)
     for (size_t i = 0; i < core_spec.size(); i++) {
-        SetRuntimeArgs(kernel, core_spec[i], runtime_args[i], blocking);
+        SetRuntimeArgsImpl(kernel, core_spec[i], runtime_args[i], blocking);
     }
 }
 
@@ -1257,7 +1280,7 @@ void DeallocateBuffer(Buffer& buffer) { buffer.deallocate(); }
 
 void AssignGlobalBufferToProgram(std::shared_ptr<Buffer> buffer, Program& program) {
     detail::DispatchStateCheck(not buffer->device()->using_slow_dispatch());
-    AddBufferToProgram(buffer, program, false);
+    program.add_buffer(buffer);
 }
 
 void SetRuntimeArgs(
